requirements.txt
numpy==1.26.4
pandas==2.2.3
smolagents[litellm,gradio]==1.9.2
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.40
langchain-google-genai==2.0.10
langchain-text-splitters==0.3.6
google-generativeai==0.8.4
faiss-cpu==1.10.0
pypdf2==3.0.1
python-dotenv==1.0.1
docling==2.25.1